+++ src/customers/customersClient.ts
@@ -1,6 +1,7 @@
 import { AbstractClient, Parameters } from '../abstractClient';
 import { GetResult } from '../getResult';
 import { DataCustomers } from './entities/dataCustomers';
+import { DataInvitation } from './entities/dataInvitation';
 
 export class CustomersClient extends AbstractClient {
   /**
@@ -14,5 +15,14 @@ export class CustomersClient extends AbstractClient {
     this.path = this.ROOT_PATH;
 
     return new GetResult(DataCustomers, await this.get(parameters));
+  }
+
+  public async getCustomerInvitation(
+    codeInvitation: string,
+    parameters: Parameters = {},
+  ): Promise<GetResult<DataInvitation>> {
+    this.path = `${this.ROOT_PATH}/invitations/${codeInvitation}`;
+
+    return new GetResult(DataInvitation, await this.get(parameters));
   }
 }