+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2022-03-21
+
+### Changed
+
+- Add endpoint get customer invitation
+
 ## [3.0.0] - 2022-03-08
 
 ### Changed